pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "debris-landlab"
version = "0.1.0"
description = "Landlab-based postfire topographic evolution and DEM differencing."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pyyaml>=6.0",
  "rasterio>=1.3",
  "fiona>=1.9",
  "geopandas>=0.14",
  "shapely>=2.0",
  "pyproj>=3.6",
  "zarr>=2.18",
  "landlab>=2.8",
  "bmi-topography>=0.5",
]

[project.optional-dependencies]
viz = [
  "matplotlib>=3.8",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[project.scripts]
dem-difference = "dem_difference:main"
export-ascii-to-tif = "export_ascii_to_tif:main"
export-raster-products = "export_ascii_to_tif:main"
resolve-workflow-config = "workflow_config:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = [
  "dem_difference",
  "export_ascii_to_tif",
  "landlab_io",
  "reproject_and_resample",
  "workflow_config",
]
